Add superconnect() to use in addition to connect()
connect() expects the exact mapping from Redux state
to props of your component when you receive it. The
problem that we have though is that TIME IS UP.

So we will send the entire redux state to every
component we superconnect, and we will assume the
component Deals™®.

Because enough of this boilerplate.