We have a problem with our auth
The way it is implemented, we are tossing currentUser from the
store on each page reload. The JWT token, meanwhile, _is_ still
present. These two things get out of sync, basically all the time,
and it is very annoying - for instance, you can still join a game
because your Feathers JWT auth is still valid, but the UI gets
confused because currentUser is null.

And apparently Feathers doesn't send all the details we need
with the JWT token it generates - it only sends us the user ID. So
ideally we need this "maintain auth", very imopure, very mutating thing
that is going to re-authenticate with feathers, then use the JWT token
it has received to perform a get() on the users service, and then stuff
the things it received into currentUser on the redux state again. Uuugh.